code cleanup and pretty-dates in the messages
the dates aren't exactly working as planned. the loops should all be
set up right, but there is some sort of problem with the comparing of
datestamps. It seems like maybe the server has the wrong time, or maybe
I'm not taking time zones in to account, or maybe I need to more
explicitly convert the timestamps so that it doesn't confuse YMD with
MYD or DMY.